safari-600.7-branch/LayoutTests/fast/ruby/ruby-text-indent.html

<head>
<script>
function test()
{
    if (window.testRunner)
        testRunner.dumpAsText();
    var testp = document.getElementById("p");
    var testr = document.getElementById("r");
    var testt = document.getElementById("t");
    var indentp = window.getComputedStyle(testp, null).getPropertyValue("text-indent");
    var indentr = window.getComputedStyle(testr, null).getPropertyValue("text-indent");
    var indentt = window.getComputedStyle(testt, null).getPropertyValue("text-indent");
    var resultp = document.getElementById("resultp");
    var resultr = document.getElementById("resultr");
    var resultt = document.getElementById("resultt");
    resultp.innerHTML = indentp;
    resultr.innerHTML = indentr;
    resultt.innerHTML = indentt;
}
</script>
</head>
<body onload="test()">
<p>This tests that ruby base and text are not affected by CSS text-indent.</p>
<p style="text-indent: 50px" id="p">This is <ruby id="r">indented by 50 pixel<rt id="t">&quot;text-indent: 50px;&quot;</rt></ruby>, but the ruby should flow without extra space.</p>
<p>text-indent for &lt;p&gt;: <span id="resultp">FAILED!</span></p>
<p>text-indent for &lt;ruby&gt;: <span id="resultr">FAILED!</span></p>
<p>text-indent for &lt;rt&gt;: <span id="resultt">FAILED!</span></p>
</body>
